handle multiple instances and unregister

+++ routes/index.js
@@ -23,7 +23,6 @@ router.all("/:apiName/:path", (req, res) => {
 
 router.post("/register", (req, res) => {
   const registrationInfo = req.body;
-
   registrationInfo.url =
     registrationInfo.protocol +
     "://" +
@@ -32,17 +31,67 @@ router.post("/register", (req, res) => {
     registrationInfo.port +
     "/";
 
-  registry.services[registrationInfo.apiName] = {
-    ...registrationInfo,
-  };
+  if (apiAlreadyExists(registrationInfo)) {
+    res.send(
+      "Config already exists for '" +
+        registrationInfo.apiName +
+        "' at '" +
+        registrationInfo.url +
+        "'"
+    );
+  } else {
+    registry.services[registrationInfo.apiName].push({
+      ...registrationInfo,
+    });
 
-  fs.writeFile("./routes/registry.json", JSON.stringify(registry), (err) => {
-    if (err) {
-      res.send("Could not register " + registrationInfo.apiName + ":" + err);
-    } else {
-      res.send("Successfully register '" + registrationInfo.apiName + "'");
+    fs.writeFile("./routes/registry.json", JSON.stringify(registry), (err) => {
+      if (err) {
+        res.send("Could not register " + registrationInfo.apiName + ":" + err);
+      } else {
+        res.send("Successfully register '" + registrationInfo.apiName + "'");
+      }
+    });
+  }
+});
+
+router.post("unregister", (req, res) => {
+  const registrationInfo = req.body;
+
+  if (apiAlreadyExists(registrationInfo)) {
+    const index = registry.services[registrationInfo.apiName].findIndex(
+      (instance) => registrationInfo.url === instance.url
+    );
+    registry.services[registrationInfo.apiName].splice(index, 1);
+
+    fs.writeFile("./routes/registry.json", JSON.stringify(registry), (err) => {
+      if (err) {
+        res.send(
+          "Could not unregister " + registrationInfo.apiName + ":" + err
+        );
+      } else {
+        res.send("Successfully unregister '" + registrationInfo.apiName + "'");
+      }
+    });
+  } else {
+    "Config does not exist for '" +
+      registrationInfo.apiName +
+      "' at '" +
+      registrationInfo.url +
+      "'";
+  }
+});
+
+function apiAlreadyExists(registrationInfo) {
+  let exists = false;
+
+  registry.services[registrationInfo.apiName].forEach((instance) => {
+    if (instance.url === registrationInfo.url) {
+      exists = true;
+      return;
     }
   });
-});
+
+  return exists;
+}
 
 module.exports = router;
